requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
ccxt==4.3.85
ta==0.11.0
tqdm==4.66.4
python-dateutil==2.9.0.post0
